formula name corrected. Test added for calculations and formula

+++ Bezier.Logic/Infrastructure/ICalculatePointUnit.cs
@@ -2,6 +2,6 @@
 {
     public interface ICalculatePointUnit
     {
-        double CalculatePointUnit(double[] pointUnits, int interval);
+        double CalculatePointUnit(double[] pointUnits, double interval);
     }
 }
